Generate integration test configs
