+++ packages/asana/svgTransform.js
@@ -6,6 +6,6 @@ module.exports = {
   },
   getCacheKey() {
     // The output is always the same.
-    return "svgTransform";
+    return 'svgTransform';
   },
 };
